Weekly Update
This week I focused on trying to get the new GDAL package to work on Windows 10 before trying to run it on Windows Server. Had a confusing issue with building the core of LANDIS-II that led me to believe it was running with the new version of GDAL but it was still using the old one for some reason. I added some debug code to Library-Spatial to help me diagnose what is going run.

5.5 hours total

1 hours researching NuGet and how its packages work within Visual Studio

1 hour creating a NuGet package and account for the new version of GDAL within Library-Spatial

3.5 hours trying to build and include this package in the core of LANDIS-II v7 and attempting to run it with the new GDAL package. Ran into the same issue as seen when running the old GDAL on Windows Server which I hope is because I configured something wrong and not that the new version of GDAL will not work. Further testing is required. I may contact the creator of the GDAL package for their help.